12/feb/2020 product work

+++ Dailyfoods/Models/category.cs
@@ -7,7 +7,7 @@ namespace Dailyfoods.Models
 {
     public class category
     {
-        public int id { get; set; }
+        public Byte id { get; set; }
         public string name { get; set; }
 
     }
